// app/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  ADMIN: 'ADMIN',
  ANALYST: 'ANALYST',
  VIEWER: 'VIEWER'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const Sentiment = {
  POS: 'POS',
  NEU: 'NEU',
  NEG: 'NEG'
} as const

export type Sentiment = (typeof Sentiment)[keyof typeof Sentiment]


export const Status = {
  NEW: 'NEW',
  REVIEWED: 'REVIEWED',
  ACTIONED: 'ACTIONED'
} as const

export type Status = (typeof Status)[keyof typeof Status]


export const FeedbackStatus = {
  PENDING: 'PENDING',
  REVIEWED: 'REVIEWED',
  ANALYZED: 'ANALYZED'
} as const

export type FeedbackStatus = (typeof FeedbackStatus)[keyof typeof FeedbackStatus]